requirements.txt
# Core ML dependencies
tensorflow>=2.15.0
tf-keras>=2.15.0
numpy>=1.24.0
Pillow>=10.0.0
scipy>=1.11.0

# Akida MetaTF packages
akida>=2.0.0
quantizeml>=1.5.0
cnn2snn>=2.0.0
akida-models>=2.0.0

# Data processing
tensorflow-datasets>=4.9.0
opencv-python>=4.8.0

# Training utilities
tensorboard>=2.15.0
scikit-learn>=1.3.0

# Configuration
pyyaml>=6.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Utilities
tqdm>=4.66.0
